+++ packages/react-core/src/core/providers/thirdweb-provider.tsx
@@ -1,5 +1,5 @@
 import { ThirdwebAuthProvider } from "../../evm/contexts/ThirdwebAuthProvider";
-import { useUpdateChainsWithApiKeys } from "../../evm/hooks/chain-hooks";
+import { useUpdateChainsWithClientId } from "../../evm/hooks/chain-hooks";
 import { ThirdwebSDKProvider } from "../../evm/providers/thirdweb-sdk-provider";
 import { ThirdwebSDKProviderProps } from "../../evm/providers/types";
 import { WalletConfig } from "../types/wallet";
@@ -106,12 +106,10 @@ export const ThirdwebProviderCore = <TChains extends Chain[]>({
   }, [props.supportedChains, activeChain]);
 
   const [supportedChainsWithKey, activeChainIdOrObjWithKey] =
-    useUpdateChainsWithApiKeys(
+    useUpdateChainsWithClientId(
       supportedChainsNonNull,
       props.activeChain || supportedChainsNonNull[0],
-      props.thirdwebApiKey,
-      props.alchemyApiKey,
-      props.infuraApiKey,
+      props.clientId,
     );
 
   const activeChainWithKey = useMemo(() => {
@@ -155,6 +153,7 @@ export const ThirdwebProviderCore = <TChains extends Chain[]>({
         activeChain={activeChainWithKey}
         autoSwitch={props.autoSwitch}
         autoConnectTimeout={props.autoConnectTimeout}
+        clientId={props.clientId}
       >
         <ThirdwebSDKProviderWrapper
           queryClient={props.queryClient}
@@ -163,9 +162,7 @@ export const ThirdwebProviderCore = <TChains extends Chain[]>({
           activeChain={activeChainWithKey}
           storageInterface={props.storageInterface}
           authConfig={props.authConfig}
-          thirdwebApiKey={props.thirdwebApiKey}
-          alchemyApiKey={props.alchemyApiKey}
-          infuraApiKey={props.infuraApiKey}
+          clientId={props.clientId}
         >
           <ThirdwebAuthProvider value={props.authConfig}>
             {props.children}