+++ packages/react-native/src/evm/providers/ui-context-provider.tsx
@@ -4,6 +4,7 @@ import { ThemeProviderProps } from "../styles/ThemeProvider";
 import { Theme, darkTheme, lightTheme } from "../styles/theme";
 import { useAppTheme } from "../styles/hooks";
 import { useTheme } from "@shopify/restyle";
+import { PropsWithChildren } from "react";
 
 type UIContextType = {
   modalState: ModalState;
@@ -25,7 +26,7 @@ const UIContext = createContext<UIContextType>({
   setTheme: () => {},
 });
 
-export const UIContextProvider = (props: React.PropsWithChildren<{}>) => {
+export const UIContextProvider = (props: PropsWithChildren<{}>) => {
   const [modalState, setModalState] = useState<ModalState>({
     view: "Closed",
     data: {},