+++ packages/react/src/wallet/wallets/metamask/MetamaskConnectUI.tsx
@@ -13,6 +13,8 @@ export const MetamaskConnectUI = (props: ConnectUIProps<MetaMaskWallet>) => {
   const { walletConfig, close } = props;
   const connect = useConnect();
 
+  const hideBackButton = props.supportedWallets.length === 1;
+
   const { goBack } = props;
 
   const connectPrompted = useRef(false);
@@ -56,6 +58,7 @@ export const MetamaskConnectUI = (props: ConnectUIProps<MetaMaskWallet>) => {
   if (screen === "connecting") {
     return (
       <ConnectingScreen
+        hideBackButton={hideBackButton}
         onBack={props.goBack}
         walletName={walletConfig.meta.name}
         walletIconURL={walletConfig.meta.iconURL}
@@ -72,7 +75,9 @@ export const MetamaskConnectUI = (props: ConnectUIProps<MetaMaskWallet>) => {
         chromeExtensionLink={walletConfig.meta.urls?.chrome}
         googlePlayStoreLink={walletConfig.meta.urls?.android}
         appleStoreLink={walletConfig.meta.urls?.ios}
-        onBack={props.goBack}
+        onBack={() => {
+          setScreen("scanning");
+        }}
       />
     );
   }
@@ -80,11 +85,12 @@ export const MetamaskConnectUI = (props: ConnectUIProps<MetaMaskWallet>) => {
   if (screen === "scanning") {
     return (
       <MetamaskScan
-        onBack={props.goBack}
+        onBack={goBack}
         onConnected={close}
         onGetStarted={() => {
           setScreen("get-started");
         }}
+        hideBackButton={hideBackButton}
         walletConfig={walletConfig}
       />
     );