+++ packages/sdk/src/evm/contracts/prebuilt-implementations/pack.ts
@@ -141,6 +141,7 @@ export class Pack extends StandardErc1155<PackContract> {
             },
           }
         : options,
+      storage,
     ),
   ) {
     super(contractWrapper, storage, chainId);
@@ -355,7 +356,7 @@ export class Pack extends StandardErc1155<PackContract> {
    *   // ERC20 rewards to be included in the pack
    *   erc20Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       quantityPerReward: 5,
    *       quantity: 100,
    *       totalRewards: 20,
@@ -364,14 +365,14 @@ export class Pack extends StandardErc1155<PackContract> {
    *   // ERC721 rewards to be included in the pack
    *   erc721Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       tokenId: 0,
    *     }
    *   ],
    *   // ERC1155 rewards to be included in the pack
    *   erc1155Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       tokenId: 0,
    *       quantityPerReward: 1,
    *       totalRewards: 100,
@@ -404,7 +405,7 @@ export class Pack extends StandardErc1155<PackContract> {
    *   // ERC20 rewards to be included in the pack
    *   erc20Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       quantityPerReward: 5,
    *       quantity: 100,
    *       totalRewards: 20,
@@ -413,14 +414,14 @@ export class Pack extends StandardErc1155<PackContract> {
    *   // ERC721 rewards to be included in the pack
    *   erc721Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       tokenId: 0,
    *     }
    *   ],
    *   // ERC1155 rewards to be included in the pack
    *   erc1155Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       tokenId: 0,
    *       quantityPerReward: 1,
    *       totalRewards: 100,
@@ -484,7 +485,7 @@ export class Pack extends StandardErc1155<PackContract> {
    *   // ERC20 rewards to be included in the pack
    *   erc20Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       quantityPerReward: 5,
    *       quantity: 100,
    *       totalRewards: 20,
@@ -493,14 +494,14 @@ export class Pack extends StandardErc1155<PackContract> {
    *   // ERC721 rewards to be included in the pack
    *   erc721Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       tokenId: 0,
    *     }
    *   ],
    *   // ERC1155 rewards to be included in the pack
    *   erc1155Rewards: [
    *     {
-   *       assetContract: "0x...",
+   *       contractAddress: "0x...",
    *       tokenId: 0,
    *       quantityPerReward: 1,
    *       totalRewards: 100,
@@ -774,7 +775,8 @@ export class Pack extends StandardErc1155<PackContract> {
    * @internal
    */
   public async prepare<
-    TMethod extends keyof PackContract["functions"] = keyof PackContract["functions"],
+    TMethod extends
+      keyof PackContract["functions"] = keyof PackContract["functions"],
   >(
     method: string & TMethod,
     args: any[] & Parameters<PackContract["functions"][TMethod]>,
@@ -792,7 +794,8 @@ export class Pack extends StandardErc1155<PackContract> {
    * @internal
    */
   public async call<
-    TMethod extends keyof PackContract["functions"] = keyof PackContract["functions"],
+    TMethod extends
+      keyof PackContract["functions"] = keyof PackContract["functions"],
   >(
     functionName: string & TMethod,
     args?: any[] & Parameters<PackContract["functions"][TMethod]>,