+++ packages/sdk/src/index.ts
@@ -1,4 +1,5 @@
 // handle browser vs node global
+
 // eslint-disable-next-line better-tree-shaking/no-top-level-side-effects
 globalThis.global = globalThis;
 
@@ -11,6 +12,8 @@ export type {
 
 export type { CurrencyValue, TokenMetadata } from "./core/schema/token";
 export { getRpcUrl } from "./core/constants/urls";
+
+export { checkClientIdOrSecretKey } from "./core/utils/apiKey";
 
 export type { QueryAllParams } from "./core/schema/QueryParams";